Fix all the code sniffer errors
